feat: mejorar validaciones y mensajes de error en la eliminación de alumnos, mesas y profesores; actualizar filtros en la vista de carreras

+++ app/Http/Controllers/Admin/MesasCrudController.php
@@ -181,7 +181,20 @@ class MesasCrudController extends BaseController
                 return redirect()->route('admin.mesa.index')
                     ->with('error', 'No se pudo eliminar la mesa. Tiene fecha hoy o futura.');
             }
+
+            //Verificar que no tenga alumnos inscriptos
+            if ($mesa->alumnos()->exists()) {
+                return redirect()->route('admin.mesa.index')
+                    ->with('error', 'No se pudo eliminar la mesa. Tiene alumnos inscriptos.');
+            }
+
+            //verificar que no tenga profesores asignados
+            if ($mesa->prof_presidente != 0 || $mesa->prof_vocal_1 != 0 || $mesa->prof_vocal_2 != 0) {
+                return redirect()->route('admin.mesa.index')
+                    ->with('error', 'No se pudo eliminar la mesa. Tiene profesores asignados.');
+            }
             
+            //eliminar mesa
             $mesa->delete();
             return redirect()->route('admin.mesa.index')
                 ->with('mensaje', 'Se ha eliminado el alumno');
